tests: refactor tests.cpp,main.cpp,LogicTest.cpp
Summary of changes:
- rm'd tests.cpp
- moved 'default test' back into main.cpp. This is the common entry point for ALPACA statement entry
- split the rest of the tests into LogicTest. You can now build + run the tests with "cmake . && make logic-test && ./logic-test"
- fixed build errors
